% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oii_association_measures.R
\name{concordant.pairs}
\alias{concordant.pairs}
\title{The number of concordant pairs in a table or matrix}
\usage{
concordant.pairs(x, y = NULL)
}
\arguments{
\item{x}{a table or matrix if \code{y} is NULL, or a numeric vector for the row variable}

\item{y}{the column variable, a numeric vector used only when \code{x} is not a table or matrix.}
}
\value{
The number of concordant pairs
}
\description{
The number of concordant pairs in a table or matrix
}
\seealso{
\code{\link{association.measures}}, \code{\link{discordant.pairs}}, \code{\link{tied.pairs}}
}

