% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oii.R
\docType{package}
\name{oii}
\alias{oii}
\alias{oii-package}
\title{Commands for the OII MSc Stats course}
\description{
This package provides a few commands that are used within the
MSc course at the Oxford Internet Institute, University of Oxford
}
\details{
The only functions you're likely to need from \pkg{oii} are
\code{\link{oii.summary}}, \code{\link{oii.freq}}, and \code{\link{oii.xtab}}.
}

