% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oii_association_measures.R
\name{tied.pairs}
\alias{tied.pairs}
\title{The number of tied pairs, a measure of association}
\usage{
tied.pairs(x, y = NULL)
}
\arguments{
\item{x}{a table or matrix if \code{y} is NULL, or a numeric vector for the first variable}

\item{y}{the second variable, a numeric vector used only when \code{x} is not a table or matrix.}
}
\value{
A list with the following values:
  \item{first}{The number of pairs tied on the first variable, but not both variables}
  \item{second}{The number of pairs tied on the second variable, but not both variables}
  \item{both}{The number of pairs tied on both the first and second variables}
}
\description{
The number of tied pairs, a measure of association
}
\seealso{
\code{\link{association.measures}}, \code{\link{concordant.pairs}}, \code{\link{discordant.pairs}}
}

